move palette ops into lib
