+++ gateware/src/rs/lib/src/lib.rs
@@ -4,3 +4,4 @@ pub mod opt;
 pub mod draw;
 pub mod log;
 pub mod generated_constants;
+pub mod palette;
